requirements.txt
Django==4.1.4
djangorestframework==3.13.1
ffprobe==0.5
future==0.18.2
huey==2.4.5
flower
coreapi==2.3.3
django-decorator-include==3.0
currencies==2020.12.12
django-redis==5.2.0
CurrencyConverter==0.17.3
Pillow==9.1.0
numpy==1.24.2
psycopg2-binary==2.9.3
drf_nested_forms==1.1.7
requests==2.27.1
urllib3==1.26.9
tuspy==0.2.5
lxml==4.9.1
django-filter==22.1
python-dateutil==2.8.2
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1
jwcrypto==1.3.1
PyYAML==6.0
python-quickbooks==0.9.2
nplusone==1.0.0
django-cors-headers==3.13.0
boto3==1.26.16
gunicorn==20.1.0
storage==0.0.4.3
django-storages==1.13.1
pyyaml==6.0
PyMuPDF==1.19.0
smart-open==6.3.0
uritemplate==4.1.1
qencode @ git+https://github.com/ManojDatt/qencode-api.git
djangorestframework-simplejwt==5.2.2
arrow==1.2.3
Werkzeug==2.2.2
drf-standardized-errors==0.12.5
redis==4.4.2
Faker==15.3.4
forex-python==1.8
whitenoise==6.4.0
